requirements.txt
langchain
transformers
llama-cpp-python
PyPDF2
tiktoken
faiss-cpu --timeout=60
fastapi
sentence_transformers
sentencepiece